Update to upstream version '2.17.58'
with Debian dir 05c22b6d4f9eff02f6c1b8917831ca07ea0c1aad

+++ lustre/llite/rw.c
@@ -1874,8 +1874,8 @@ int ll_readpage(struct file *file, struct page *vmpage)
 	 * the PCC copy.
 	 */
 	if (inode->i_op != &ll_file_inode_operations) {
-		CERROR("%s: readpage() on invalidated PCC inode %lu: rc=%d\n",
-		       sb->s_id, inode->i_ino, -EIO);
+		CERROR("%s: readpage() on invalidated PCC inode %llu: rc=%d\n",
+		       sb->s_id, (u64)inode->i_ino, -EIO);
 		unlock_page(vmpage);
 		RETURN(-EIO);
 	}
@@ -1994,7 +1994,7 @@ int ll_readpage(struct file *file, struct page *vmpage)
 		 * Because this page is beyond the boundary of the requested
 		 * read, Lustre does not lock it as part of the read.  This
 		 * means we must check if there is a valid dlmlock on this
-		 * this page and reference it before we attempt to read in the
+		 * page and reference it before we attempt to read in the
 		 * page.  If there is not a valid dlmlock, then we are racing
 		 * with dlmlock cancellation and the page is being removed
 		 * from the cache.
